decomp progress: 3 big structural wins
- WiiContent::PopAfterRestore 2.86%->99.54% (full body + int return)
- StoreRbnOfferTable::Load 1.37%->99.7% (sibling-pattern body)
- CharIKFingers::CalculateHandDest 60.22%->76.76% ((float) cast)

+++ src/system/char/CharIKFingers.cpp
@@ -279,7 +279,7 @@ void CharIKFingers::CalculateHandDest(int i1, int i2) {
                 if (curDesc.unk0) {
                     ::Add(curDesc.unk8, v194, v194);
                     Vector3 v1ac;
-                    Scale(v1a0, i - 2.0, v1ac);
+                    Scale(v1a0, (float)(i - 2), v1ac);
                     ::Add(v1ac, v194, v194);
                     if (i == 0) {
                         Hmx::Matrix3 m158;
@@ -300,7 +300,7 @@ void CharIKFingers::CalculateHandDest(int i1, int i2) {
                     }
                 }
             }
-            Scale(v194, 1.0f / i1, v194);
+            Scale(v194, 1.0f / (float)i1, v194);
             if (b1)
                 v188.y += mHandMoveForward;
             ::Add(mHandKeyboardOffset, v188, v188);
